backend/requirements.txt
# -----------------------------
# Web framework & API
# -----------------------------
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# -----------------------------
# AWS SDK (Bedrock)
# -----------------------------
boto3>=1.34.0
botocore>=1.34.0

# -----------------------------
# Data validation
# -----------------------------
pydantic>=2.6.0

# -----------------------------
# Prompt loading (YAML)
# -----------------------------
pyyaml>=6.0.0

# -----------------------------
# Local development only
# -----------------------------
python-dotenv>=1.0.0
